pyproject.toml
[project]
name = "funnel-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agentql>=1.18.1",
    "fpdf2>=2.8.7",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
